Combine selection actions
Previous to this commit, there would be weird bugs with actions inside
ListFragments originating from onClick or onMouseDown handlers. Two
actions were issued at the same time, which would cause the app state to
be read too early by the second reducer.

Obviously using Redux would fix this. However, I don't yet know Redux.
The plan is to make this app run, then make it good. Since this app's
architecture is nearly homomorphic to that of Redux, I should be able to
just slap it on once the app is more complete.